+++ Templates/scripts/transfer-issue.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-// **Version:** 0.15.2
+// **Version:** 0.15.3
 /**
  * transfer-issue.js
  *